requirements.txt
accelerate==1.3.0
attrdict==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
einops==0.8.0
filelock==3.13.1
flatbuffers==25.1.24
fsspec==2024.6.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
-e git+https://github.com/deepseek-ai/Janus.git@1daa72fa409002d40931bd7b36a9280362469ead#egg=janus
Jinja2==3.1.4
MarkupSafe==2.1.5
moondream==0.0.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
protobuf==5.29.3
psutil==6.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
